renaming script_libs to ops
