+++ src/vtp/ops/accept_ballot_lib.py
