+++ src/vtp/ops/show_contests_lib.py
