+++ src/vtp/ops/tally_contests_lib.py
