+++ src/vtp/ops/verify_ballot_receipt_lib.py
