+++ src/vtp/setup_vtp_demo.py
@@ -27,7 +27,7 @@ See 'setup_vtp_demo -h' for usage information.
 # pylint: disable=wrong-import-position   # import statements not top of file
 import sys
 
-from vtp.script_libs.setup_vtp_demo_lib import SetupVtpDemoLib
+from vtp.ops.setup_vtp_demo_lib import SetupVtpDemoLib
 from vtp.utils.election_config import ElectionConfig
 
 
@@ -53,4 +53,4 @@ def main():
 
 # If called directly via this file
 if __name__ == "__main__":
-    main
+    main()
