+++ src/vtp/show_contest.py
@@ -26,7 +26,7 @@ See 'show_contest.py -h' for usage information.
 # pylint: disable=wrong-import-position   # import statements not top of file
 import sys
 
-from vtp.script_libs.show_contests_lib import ShowContestsLib
+from vtp.ops.show_contests_lib import ShowContestsLib
 from vtp.utils.election_config import ElectionConfig
